backend/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.8.2
Pillow>=10.0.0
python-multipart>=0.0.5 